+++ examples/docs/task/output/pass-stdout.ts
@@ -1,4 +1,4 @@
-import { execaCommand as command } from 'execa'
+import { execa } from 'execa'
 
 import { Listr } from 'listr2'
 
@@ -6,7 +6,7 @@ const tasks = new Listr([
   {
     title: 'This task will do a curl request.',
     task: async(_, task): Promise<void> => {
-      const execute = command('curl -v http://google.com')
+      const execute = execa('curl', ['-v', 'http://google.com'])
 
       execute.stdout.pipe(task.stdout())
       execute.stderr.pipe(task.stdout())